*Skip spaces at the end of a category in WiiTDB (e.g 'Rhythm  ') *Change in nand title search (fixed bug which can find wrong nand title and set it as return to) *Fixed bug when a wiimote leaves the game list choice and returns to it without going over another game that game was not clickable (Issue 1918) *Changed search bar to clear text on [X] button click. search bar is now closed by pressing B

+++ source/wad/nandtitle.h
@@ -88,15 +88,12 @@ class NandTitle
         bool Exists(u64 tid);
         bool ExistsFromIndex(u32 i);
 
-        int FindU64(const char *s);
-        int FindU32(const char *s);
+        u64 FindU64(const char *s);
+        u64 FindU32(const char *s);
 
         s32 GetTicketViews(u64 tid, tikview **outbuf, u32 *outlen);
 
-        u64 operator[](u32 i)
-        {
-            return At(i);
-        }
+        u64 operator[](u32 i) { return At(i); }
 
         static int LoadFileFromNand(const char *filepath, u8 **outbuffer, u32 *outfilesize);
     private:
